Compare keys by equality identity in `IndexedDict._pop_key`
There is no guarantee that the given key and the key in the dictionary will be
identical.  There is only a guarantee that they will compare equal to each
other.

Something has changed on CPython 3.11 so that even the very simple case of a
short string key the underlying objects are not necessarily identical
anymore (though in previous CPython versions they typically would be).